+++ src/main/java/fr/robotv2/betterdailyquest/listeners/StatisticListeners.java
@@ -1,6 +1,9 @@
 package fr.robotv2.betterdailyquest.listeners;
 
 import fr.robotv2.betterdailyquest.BetterDailyQuest;
+import fr.robotv2.betterdailyquest.event.PlayerBoatEvent;
+import fr.robotv2.betterdailyquest.event.PlayerJumpEvent;
+import fr.robotv2.betterdailyquest.event.PlayerMinecartEvent;
 import fr.robotv2.betterdailyquest.event.PlayerSwimEvent;
 import fr.robotv2.betterdailyquest.event.PlayerWalkEvent;
 import fr.robotv2.betterdailyquest.util.stats.StatisticWatcher;
@@ -17,12 +20,7 @@ import java.util.List;
 public class StatisticListeners implements Listener {
 
     public static final int DELAY = 20;
-    private final List<StatisticWatcher> watchers = List.of(
-            new StatisticWatcher(Statistic.WALK_ONE_CM, PlayerWalkEvent::new),
-            new StatisticWatcher(Statistic.SPRINT_ONE_CM, PlayerWalkEvent::new),
-            new StatisticWatcher(Statistic.CROUCH_ONE_CM, PlayerWalkEvent::new),
-            new StatisticWatcher(Statistic.SWIM_ONE_CM, PlayerSwimEvent::new)
-    );
+    private final List<StatisticWatcher> watchers = createWatchers();
 
     public StatisticListeners() {
         Bukkit.getOnlinePlayers().forEach(this::initialize);
@@ -52,5 +50,17 @@ public class StatisticListeners implements Listener {
         for(StatisticWatcher watcher : watchers) {
             watcher.init(player);
         }
+    }
+
+    static List<StatisticWatcher> createWatchers() {
+        return List.of(
+                new StatisticWatcher(Statistic.WALK_ONE_CM, PlayerWalkEvent::new),
+                new StatisticWatcher(Statistic.SPRINT_ONE_CM, PlayerWalkEvent::new),
+                new StatisticWatcher(Statistic.CROUCH_ONE_CM, PlayerWalkEvent::new),
+                new StatisticWatcher(Statistic.SWIM_ONE_CM, PlayerSwimEvent::new),
+                new StatisticWatcher(Statistic.JUMP, PlayerJumpEvent::new),
+                new StatisticWatcher(Statistic.BOAT_ONE_CM, PlayerBoatEvent::new),
+                new StatisticWatcher(Statistic.MINECART_ONE_CM, PlayerMinecartEvent::new)
+        );
     }
 }